+++ core/admin/PostBack.php
@@ -4,7 +4,7 @@ namespace Dev4Press\Plugin\DemoPress\Admin;
 
 use Dev4Press\v50\Core\Admin\PostBack as BasePostBack;
 use Dev4Press\v50\Core\Options\Process;
-use function Dev4Press\v50\Functions\sanitize_key_expanded;
+use Dev4Press\v50\Core\Quick\Sanitize;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -86,7 +86,7 @@ class PostBack extends BasePostBack {
 		$input = isset( $_POST['demopress_value'] ) ? (array) $_POST['demopress_value'] : array();
 
 		if ( ! empty( $input ) ) {
-			$gen_input = isset( $input['demo-generator-type'] ) ? sanitize_key_expanded( $input['demo-generator-type'] ) : '';
+			$gen_input = isset( $input['demo-generator-type'] ) ? Sanitize::key( $input['demo-generator-type'] ) : '';
 			$generator = demopress()->get_generator( $gen_input );
 
 			if ( ! is_wp_error( $generator ) ) {