+++ forms/content-about-minor.php
@@ -1,3 +1,13 @@
+<?php
+
+use Dev4Press\v50\Core\Quick\KSES;
+use function Dev4Press\v50\Functions\panel;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+?>
 <div class="d4p-about-minor">
     <h3><?php _e( "Maintenance and Security Releases", "demopress" ); ?></h3>
     <p>
@@ -5,6 +15,11 @@
         New builders. Few fixes.
     </p>
     <p>
-		<?php printf( __( "For more information, see <a href='%s'>the changelog</a>.", "demopress" ), 'admin.php?page=demopress&panel=about&subpanel=changelog' ); ?>
+		<?php
+
+		/* translators: Changelog subpanel information. %s: Subpanel URL. */
+		echo KSES::standard( sprintf( __( 'For more information, see <a href=\'%s\'>the changelog</a>.', 'coreactivity' ), esc_url( panel()->a()->panel_url( 'about', 'changelog' ) ) ) ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+
+		?>
     </p>
 </div>