+++ forms/content-tools-updater.php
@@ -1,6 +1,13 @@
+<?php
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+?>
 <div class="d4p-content">
     <div class="d4p-group d4p-group-information d4p-group-updater">
-        <h3><?php _e( "Update status", "demopress" ); ?></h3>
+    <h3><?php _e( "Update status", "demopress" ); ?></h3>
         <div class="d4p-group-inner">
             <div>
 				<?php _e( "There is nothing available for update.", "demopress" ); ?>